Move FbxLayerElementTemplate to its own file.
