+++ FbxSharp.csproj
@@ -135,6 +135,7 @@
     <Compile Include="FbxLayerElementNormal.cs" />
     <Compile Include="FbxLayerElementPolygonGroup.cs" />
     <Compile Include="FbxLayerElementSmoothing.cs" />
+    <Compile Include="FbxLayerElementTemplate.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
